requirements.txt
python-telegram-bot==21.1
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
sqlite3
asyncio
